# [swift-4.0-branch][stdlib] Remove the Sequence requirement from BinaryInteger.Words