@moiseev 2017-09-01
@swift-ci Please benchmark