@moiseev 2017-09-01
Oh well... It's failing on 32-bit platform.