@moiseev 2017-09-06
@swift-ci Please test