@graydon 2017-09-06
Good news: this appears to win back a decent chunk of frontend workload:

207292	51958	-74.93	Generic signature builder.NumConcreteTypeConstraints
5454274	5117715	-6.17	Generic signature builder.NumConformanceConstraints
4227112	3911953	-7.46	Generic signature builder.NumConformances
3114460	2307835	-25.9	Generic signature builder.NumPotentialArchetypes
3032942	2219424	-26.82	Generic signature builder.NumSameTypeConstraints
541162	498839	-7.82	Sema.NumConformancesDeserialized
2010082	1915338	-4.71	Sema.NumDeclsDeserialized
400494	379623	-5.21	Sema.NumLazyGenericEnvironments
1745340	1705310	-2.29	Sema.NumTypesDeserialized
2010082	1915338	-4.71	Serialization.NumDeclsLoaded
400494	379623	-5.21	Serialization.NumLazyGenericEnvironments
13426	11750	-12.48	Serialization.NumNormalProtocolConformancesCompleted
76440	72910	-4.62	Serialization.NumNormalProtocolConformancesLoaded

Bad news: seems to increase backend workload & codesize:

71765	85085	18.56	IRModule.NumIRBasicBlocks
758582	789837	4.12	IRModule.NumIRInsts
17123328	17731808	3.55	LLVM.NumLLVMBytesOutput

Overall looks to knock about 2s off a 60s wall time (on 8-way machine), or about 8s off 216s user time. Not a huge win, but it's significant.