@moiseev 2017-09-06
@graydon thanks! I think I can do something about the basic blocks.