@moiseev 2017-09-07
@graydon with the latest change (some reshuffling of `@_transparent` and `@inline(__always)` I get the following:

old | new | delta_pct | name
-- | -- | -- | --
25615 | 4492 | -82.46 | Generic signature builder.NumConcreteTypeConstraints
11620 | 4165 | -64.16 | dagcombine.NodesCombined
577051 | 285079 | -50.6 | time.swift.Type checking / Semantic analysis.sys
641836 | 383374 | -40.27 | Generic signature builder.NumSameTypeConstraints
255 | 170 | -33.33 | AST.NumDependencies
51 | 34 | -33.33 | IRModule.NumIRNamedMetaData
427746 | 297766 | -30.39 | Generic signature builder.NumPotentialArchetypes
1062244 | 760932 | -28.37 | Generic signature builder.NumConformanceConstraints
842577 | 604282 | -28.28 | time.swift.AST verification.wall
823510 | 613501 | -25.5 | time.swift.SIL verification, pre-optimization.wall
758057 | 566175 | -25.31 | time.swift.AST verification.user
767168 | 590264 | -23.06 | time.swift.SIL verification, pre-optimization.user
5316431 | 4131869 | -22.28 | time.swift.Type checking / Semantic analysis.wall
 |  |  | 
14967 | 15078 | 0.74 | IRModule.NumIRBasicBlocks
138307 | 144585 | 4.54 | IRModule.NumIRInsts


And the overall `time`:

what | real | user | sys
-- | -- | -- | --
old |	0m15.245s | 0m12.980s | 0m2.130s
new | 0m12.872s | 0m11.128s | 0m1.618s

So about ~18% improvement.