@graydon 2017-09-07
Interesting. It actually seems to perform a bit _worse_ than the previous iteration of this against RxSwift. I'll try reconfirming (will take a while tho).