@moiseev 2017-09-07
I've also ran benchmarks, and there are regressions ;-(

<details open>
  <summary>Regression (39)</summary>
  
TEST                                                      | OLD    | NEW    | DELTA         | SPEEDUP     
---                                                       | ---    | ---    | ---           | ---         
ArrayLiteral                                              | 0      | 1040   | +104000000.0% | **0.00x**   
OpenClose                                                 | 5      | 62     | +1139.8%      | **0.08x**   
StaticArray                                               | 6      | 16     | +166.6%       | **0.38x**   
Chars                                                     | 480    | 1157   | +141.0%       | **0.41x**   
NSStringConversion                                        | 360    | 867    | +140.8%       | **0.42x**   
ObjectiveCBridgeStubFromNSString                          | 559    | 1110   | +98.6%        | **0.50x**   
StringHasPrefix                                           | 15     | 28     | +86.7%        | **0.54x**   
StringHasSuffix                                           | 15     | 28     | +86.7%        | **0.54x**   
ObjectiveCBridgeFromNSString                              | 924    | 1583   | +71.3%        | **0.58x**   
ObjectiveCBridgeFromNSArrayAnyObjectToStringForced        | 31919  | 50958  | +59.6%        | **0.63x**   
ObjectiveCBridgeStubFromArrayOfNSString                   | 26356  | 39889  | +51.3%        | **0.66x**   
ObjectiveCBridgeFromNSArrayAnyObjectToString              | 39516  | 54285  | +37.4%        | **0.73x**   
ObjectiveCBridgeFromNSStringForced                        | 1928   | 2613   | +35.5%        | **0.74x**   
ErrorHandling                                             | 1848   | 2485   | +34.5%        | **0.74x**   
ObjectiveCBridgeStubFromNSDate                            | 3687   | 4661   | +26.4%        | **0.79x**   
CaptureProp                                               | 4438   | 5507   | +24.1%        | **0.81x**   
SetUnion_OfObjects                                        | 5919   | 7293   | +23.2%        | **0.81x**   
ObjectiveCBridgeFromNSArrayAnyObjectForced                | 3805   | 4617   | +21.3%        | **0.82x**   
DropWhileSequenceLazy                                     | 80     | 96     | +20.0%        | **0.83x**   
SetExclusiveOr_OfObjects                                  | 7117   | 8482   | +19.2%        | **0.84x**   
Array2D                                                   | 2314   | 2703   | +16.8%        | **0.86x**   
MapReduceLazyCollection                                   | 12     | 14     | +16.7%        | **0.86x**   
DropWhileAnySeqCntRangeLazy                               | 134    | 156    | +16.4%        | **0.86x**   
DropWhileAnySeqCRangeIterLazy                             | 134    | 156    | +16.4%        | **0.86x**   
MapReduceLazyCollectionShort                              | 31     | 36     | +16.1%        | **0.86x**   
StringEqualPointerComparison                              | 292    | 339    | +16.1%        | **0.86x**   
DropWhileAnyCollectionLazy                                | 134    | 155    | +15.7%        | **0.86x**   
ObjectiveCBridgeFromNSDictionaryAnyObjectToString         | 118287 | 135829 | +14.8%        | **0.87x**   
ObjectiveCBridgeFromNSArrayAnyObject                      | 22735  | 25958  | +14.2%        | **0.88x**   
StringFromLongWholeSubstring                              | 165    | 186    | +12.7%        | **0.89x**   
ObjectiveCBridgeFromNSSetAnyObjectToString                | 73830  | 81670  | +10.6%        | **0.90x**   
ObjectiveCBridgeFromNSSetAnyObjectToStringForced          | 71908  | 79262  | +10.2%        | **0.91x**   
PopFrontUnsafePointer                                     | 7771   | 8514   | +9.6%         | **0.91x**   
ObjectiveCBridgeStubNSDateRefAccess                       | 304    | 330    | +8.6%         | **0.92x**   
AngryPhonebook                                            | 2564   | 2755   | +7.4%         | **0.93x (?)** 
FilterEvenUsingReduceInto                                 | 126    | 135    | +7.1%         | **0.93x**   
DictionarySwap                                            | 406    | 434    | +6.9%         | **0.94x**   
FilterEvenUsingReduce                                     | 1117   | 1182   | +5.8%         | **0.95x**   
FrequenciesUsingReduce                                    | 6401   | 6746   | +5.4%         | **0.95x**   

</details>

<details >
  <summary>Improvement (70)</summary>
  
TEST                                                      | OLD    | NEW    | DELTA         | SPEEDUP     
---                                                       | ---    | ---    | ---           | ---         
StringAdder                                               | 3166   | 0      | -100.0%       | **3166001.00x** 
ByteSwap                                                  | 91     | 0      | -100.0%       | **91001.00x** 
BitCount                                                  | 193    | 18     | -90.7%        | **10.72x**  
Sim2DArray                                                | 551    | 243    | -55.9%        | **2.27x**   
ArrayAppendAscii                                          | 20264  | 9638   | -52.4%        | **2.10x**   
DictionaryGroup                                           | 235    | 123    | -47.7%        | **1.91x**   
StrComplexWalk                                            | 1810   | 1001   | -44.7%        | **1.81x**   
DropLastAnyCollectionLazy                                 | 23219  | 13054  | -43.8%        | **1.78x**   
SuffixAnyCollectionLazy                                   | 23187  | 13055  | -43.7%        | **1.78x**   
PrefixAnyCollectionLazy                                   | 69664  | 40578  | -41.8%        | **1.72x**   
DropFirstAnyCollectionLazy                                | 69210  | 40770  | -41.1%        | **1.70x**   
DropLastArray                                             | 16     | 10     | -37.5%        | **1.60x**   
SuffixArray                                               | 16     | 10     | -37.5%        | **1.60x**   
ArrayAppendReserved                                       | 693    | 459    | -33.8%        | **1.51x**   
ArrayPlusEqualSingleElementCollection                     | 915    | 680    | -25.7%        | **1.35x**   
ArrayAppend                                               | 919    | 686    | -25.4%        | **1.34x**   
ArrayAppendLatin1                                         | 38272  | 28669  | -25.1%        | **1.33x**   
MapReduceAnyCollection                                    | 390    | 295    | -24.4%        | **1.32x**   
DropWhileArray                                            | 63     | 48     | -23.8%        | **1.31x**   
ArrayAppendUTF16                                          | 38607  | 29450  | -23.7%        | **1.31x**   
ObjectiveCBridgeFromNSSetAnyObjectForced                  | 5175   | 4041   | -21.9%        | **1.28x**   
PrefixWhileCountableRange                                 | 42     | 33     | -21.4%        | **1.27x**   
ReversedArray                                             | 52     | 41     | -21.2%        | **1.27x**   
ObjectiveCBridgeFromNSDictionaryAnyObjectForced           | 8024   | 6397   | -20.3%        | **1.25x**   
SuffixCountableRangeLazy                                  | 10     | 8      | -20.0%        | **1.25x**   
SuffixCountableRange                                      | 10     | 8      | -20.0%        | **1.25x**   
DropLastCountableRangeLazy                                | 10     | 8      | -20.0%        | **1.25x**   
DropLastCountableRange                                    | 10     | 8      | -20.0%        | **1.25x**   
ObjectiveCBridgeStubFromNSDateRef                         | 4772   | 3823   | -19.9%        | **1.25x**   
DropWhileAnySeqCRangeIter                                 | 23767  | 19149  | -19.4%        | **1.24x**   
MapReduceClass                                            | 3238   | 2647   | -18.3%        | **1.22x**   
MapReduce                                                 | 388    | 319    | -17.8%        | **1.22x**   
Walsh                                                     | 412    | 346    | -16.0%        | **1.19x**   
PrefixWhileArray                                          | 95     | 80     | -15.8%        | **1.19x**   
DropFirstCountableRangeLazy                               | 32     | 27     | -15.6%        | **1.19x**   
DropFirstCountableRange                                   | 32     | 27     | -15.6%        | **1.19x**   
PrefixCountableRangeLazy                                  | 32     | 27     | -15.6%        | **1.19x**   
PrefixCountableRange                                      | 32     | 27     | -15.6%        | **1.19x**   
PrefixWhileCountableRangeLazy                             | 32     | 27     | -15.6%        | **1.19x**   
DeadArray                                                 | 182    | 156    | -14.3%        | **1.17x**   
MapReduceClassShort                                       | 4594   | 3954   | -13.9%        | **1.16x**   
RangeAssignment                                           | 361    | 311    | -13.9%        | **1.16x**   
PrefixWhileSequence                                       | 356    | 309    | -13.2%        | **1.15x**   
PrefixWhileAnySeqCntRange                                 | 129    | 113    | -12.4%        | **1.14x**   
StringBuilder                                             | 1012   | 892    | -11.9%        | **1.13x**   
PrefixWhileAnyCollection                                  | 133    | 118    | -11.3%        | **1.13x**   
ObjectiveCBridgeStubURLAppendPathRef                      | 338259 | 303804 | -10.2%        | **1.11x**   
ObjectiveCBridgeStubURLAppendPath                         | 338520 | 305056 | -9.9%         | **1.11x**   
CStringLongNonAscii                                       | 2173   | 1974   | -9.2%         | **1.10x**   
DropWhileAnySequence                                      | 6510   | 5921   | -9.0%         | **1.10x**   
ProtocolDispatch                                          | 2871   | 2612   | -9.0%         | **1.10x**   
SortLargeExistentials                                     | 7622   | 6993   | -8.3%         | **1.09x**   
DictionaryGroupOfObjects                                  | 1559   | 1437   | -7.8%         | **1.08x**   
DictionarySwapOfObjects                                   | 6494   | 6038   | -7.0%         | **1.08x**   
CStringLongAscii                                          | 5071   | 4719   | -6.9%         | **1.07x**   
ArrayPlusEqualFiveElementCollection                       | 5072   | 4727   | -6.8%         | **1.07x**   
ObjectiveCBridgeToNSDictionary                            | 57257  | 53564  | -6.4%         | **1.07x**   
NopDeinit                                                 | 21020  | 19691  | -6.3%         | **1.07x**   
StringWithCString                                         | 61310  | 57576  | -6.1%         | **1.06x (?)** 
IterateData                                               | 705    | 663    | -6.0%         | **1.06x (?)** 
DropWhileAnySeqCntRange                                   | 86     | 81     | -5.8%         | **1.06x**   
StringMatch                                               | 7690   | 7245   | -5.8%         | **1.06x (?)** 
DropFirstAnySeqCRangeIterLazy                             | 25074  | 23695  | -5.5%         | **1.06x**   
PrefixAnySeqCRangeIter                                    | 19615  | 18548  | -5.4%         | **1.06x**   
DropFirstAnySeqCRangeIter                                 | 25095  | 23730  | -5.4%         | **1.06x**   
SubstringFromLongStringGeneric                            | 56     | 53     | -5.4%         | **1.06x**   
PrefixAnySeqCRangeIterLazy                                | 19590  | 18563  | -5.2%         | **1.06x**   
Integrate                                                 | 273    | 259    | -5.1%         | **1.05x**   
DropLastSequenceLazy                                      | 629    | 598    | -4.9%         | **1.05x**   
DropLastSequence                                          | 630    | 599    | -4.9%         | **1.05x**   

</details>

<details open>
  <summary>Added (1)</summary>
  
TEST                                                      | MIN    | MAX    | MEAN          | MAX_RSS     
---                                                       | ---    | ---    | ---           | ---         
GlobalClass                                               | 0      | 0      | 0             | -           

</details>

<details open>
  <summary>Removed (39)</summary>
  
TEST                                                      | MIN    | MAX    | MEAN          | MAX_RSS     
---                                                       | ---    | ---    | ---           | ---         
CharIndexing_ascii_unicodeScalars                         | 13797  | 13952  | 13861         | -           
CharIndexing_ascii_unicodeScalars_Backwards               | 10703  | 10897  | 10755         | -           
CharIndexing_chinese_unicodeScalars                       | 10415  | 10504  | 10454         | -           
CharIndexing_chinese_unicodeScalars_Backwards             | 8146   | 8200   | 8164          | -           
CharIndexing_japanese_unicodeScalars                      | 16484  | 16745  | 16556         | -           
CharIndexing_japanese_unicodeScalars_Backwards            | 12811  | 12890  | 12844         | -           
CharIndexing_korean_unicodeScalars                        | 13359  | 13586  | 13418         | -           
CharIndexing_korean_unicodeScalars_Backwards              | 10386  | 10432  | 10404         | -           
CharIndexing_punctuatedJapanese_unicodeScalars            | 2497   | 2510   | 2502          | -           
CharIndexing_punctuatedJapanese_unicodeScalars_Backwards  | 1971   | 1991   | 1979          | -           
CharIndexing_punctuated_unicodeScalars                    | 3124   | 3149   | 3133          | -           
CharIndexing_punctuated_unicodeScalars_Backwards          | 2458   | 2850   | 2562          | -           
CharIndexing_russian_unicodeScalars                       | 11492  | 11776  | 11568         | -           
CharIndexing_russian_unicodeScalars_Backwards             | 8933   | 9007   | 8967          | -           
CharIndexing_tweet_unicodeScalars                         | 27200  | 27533  | 27302         | -           
CharIndexing_tweet_unicodeScalars_Backwards               | 21488  | 21705  | 21544         | -           
CharIndexing_utf16_unicodeScalars                         | 75286  | 75727  | 75509         | -           
CharIndexing_utf16_unicodeScalars_Backwards               | 66912  | 67592  | 67160         | -           
CharIteration_ascii_unicodeScalars                        | 14955  | 15068  | 14991         | -           
CharIteration_ascii_unicodeScalars_Backwards              | 15605  | 15657  | 15644         | -           
CharIteration_chinese_unicodeScalars                      | 11307  | 11373  | 11335         | -           
CharIteration_chinese_unicodeScalars_Backwards            | 11866  | 11891  | 11878         | -           
CharIteration_japanese_unicodeScalars                     | 17925  | 17998  | 17962         | -           
CharIteration_japanese_unicodeScalars_Backwards           | 18688  | 18805  | 18744         | -           
CharIteration_korean_unicodeScalars                       | 14545  | 14865  | 14707         | -           
CharIteration_korean_unicodeScalars_Backwards             | 15447  | 15641  | 15513         | -           
CharIteration_punctuatedJapanese_unicodeScalars           | 2720   | 2784   | 2753          | -           
CharIteration_punctuatedJapanese_unicodeScalars_Backwards | 2826   | 2890   | 2858          | -           
CharIteration_punctuated_unicodeScalars                   | 3393   | 3936   | 3542          | -           
CharIteration_punctuated_unicodeScalars_Backwards         | 3520   | 3559   | 3535          | -           
CharIteration_russian_unicodeScalars                      | 12466  | 12550  | 12501         | -           
CharIteration_russian_unicodeScalars_Backwards            | 12996  | 13027  | 13015         | -           
CharIteration_tweet_unicodeScalars                        | 29548  | 29828  | 29617         | -           
CharIteration_tweet_unicodeScalars_Backwards              | 30942  | 31366  | 31120         | -           
CharIteration_utf16_unicodeScalars                        | 71594  | 71919  | 71740         | -           
CharIteration_utf16_unicodeScalars_Backwards              | 102163 | 102604 | 102396        | -           
accessGlobal                                              | 5      | 5      | 5             | -           
accessInMatSet                                            | 15     | 19     | 16            | -           
accessIndependent                                         | 2      | 2      | 2             | -           

</details>